Main Menu and Pause Screen added
Main menu scene has been added along with Pause canvas which triggers when player clicks Pause button on top of screen. All 3 scenes are connected together with help of buttons. Although, when on Pause scene and we wish to resume game, dart is being thrown towards rotator. Need to fix it...