AudioManager Implemented
To handle audio in the game, I have created an AudioManager gameobject. Where we can set up volume, pitch, looping, etc properties for the particualr audio effect. Also made its prefab and we can use it in different scenes also.
AudioManager class is used as singleton so that the music won't get cut off when scnes change and also it avoids repetition of AudioManager gameobject in any particular scene.
Feeling so proud of myself today :))